Create addalias
Find the filename of the markdown files in the folder, convert it to lower case, add a - instead of spaces between words int he filename, and add the filename as a url parameter pathname prefixing it with a /en. The url parameter should have just /en/<filename>. For example: /en/posture-host-analyzer.